pyproject.toml
[tool.poetry]
name = "Python binary auto release"
version = "0.0.1"
description = "Nuitka を使って Python スクリプトを実行ファイルに変換する GitHub Actions のサンプルです．"
authors = ["KIMATA Tetsuya <kimata@green-rabbit.net>"]


[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
docopt = "^0.6.2"

[tool.poetry.group.dev.dependencies]
nuitka = "^2.1.3"

[tool.pytest.ini_options]
minversion = "6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
